LayoutTests/media/video-loopend.html

<video src=content/test.mp4 playcount=2 loopend=0.5s controls></video>
<p>Test that video loops when time reaches the value specified by the loopEnd attribute.</p>
<script src=video-test.js></script>
<script>
    testExpected("video.loopEnd", 0.5);
    setTimeout(function () { testExpected("video.currentTime", 0.5, "<="); }, 2000);
    waitForEvent('timeupdate', function () {
        if (video.currentLoop == 1) {
            testExpected("video.currentLoop", 1);
            testExpected("video.currentTime.toFixed(1)", 0);
            endTest();
        }
    });
    run("video.play()");
</script>
