LayoutTests/media/video-loopend-expected.txt

Test that video loops when time reaches the value specified by the loopEnd attribute.

TEST(video.loopEnd==0.5) OK
RUN(video.play())
EVENT(timeupdate)
EVENT(timeupdate)
TEST(video.currentLoop == 1) OK
TEST(video.currentTime.toFixed(1) == 0) OK
END OF TEST
